+++ foxes/variables.py
@@ -104,8 +104,8 @@ P = "P"
 :group: foxes.variables
 """
 
-P_NOMINAL = "P_nominal"
-""" The nominal power
+CAP = "CAP"
+""" The capacity (equals P_nominal for wind turbines)
 :group: foxes.variables
 """
 
@@ -139,8 +139,8 @@ EFF = "EFF"
 :group: foxes.variables
 """
 
-CAP = "CAP"
-""" Capacity, equals P/P_nominal
+CAPF = "CAPF"
+""" Capacity factor, equals P/CAP
 :group: foxes.variables
 """
 
@@ -256,8 +256,8 @@ AMB_YLD = "AMB_YLD"
 :group: foxes.variables
 """
 
-AMB_CAP = "AMB_CAP"
-""" Ambient capacity, equals AMB_P/P_nominal
+AMB_CAPF = "AMB_CAPF"
+""" Ambient capacity, equals AMB_P/CAP
 :group: foxes.variables
 """
 
@@ -315,7 +315,7 @@ var2amb = {
         WEIBULL_A,
         WEIBULL_k,
         YLD,
-        CAP,
+        CAPF,
         FLF,
         UV,
         U,
@@ -340,8 +340,8 @@ extensive_farm = set(
         P,
         AMB_YLD,
         YLD,
-        AMB_CAP,
-        CAP,
+        AMB_CAPF,
+        CAPF,
     ]
 )
 """ Set of extensive variables, i.e. variables that should be summed when aggregating over turbines
@@ -362,8 +362,8 @@ extensive_state = set(
         P,
         AMB_YLD,
         YLD,
-        AMB_CAP,
-        CAP,
+        AMB_CAPF,
+        CAPF,
         WEIGHT,
     ]
 )
@@ -458,7 +458,7 @@ ROUND_DIGITS = {
     CT: 6,
     T: 3,
     YLD: 3,
-    CAP: 5,
+    CAPF: 5,
     EFF: 5,
     WEIBULL_A: 3,
     WEIBULL_k: 3,